+++ tests/testthat/test-y_documents.R
@@ -3,7 +3,7 @@ test_that("list_documents() returns document ids", {
   expect_identical(list_documents(docs), character())
 
   docs <- mighty_documents(
-    file = test_path("test_study/documents.yml")
+    file = test_path("test_study/_documents.yml")
   )
   expect_identical(list_documents(docs), c("SUPPDOC001", "COMMENT001", "METHOD001"))
 })
@@ -29,7 +29,7 @@ test_that("document can be added via add_document()", {
 
 test_that("document can be selected via select_document()", {
   docs <- mighty_documents(
-    file = test_path("test_study/documents.yml")
+    file = test_path("test_study/_documents.yml")
   )
 
   expect_equal(select_document(docs, "SUPPDOC001")$title, "Analysis Reviewer Guide Protocol")
@@ -37,7 +37,7 @@ test_that("document can be selected via select_document()", {
 
 test_that("document title can be correctly updated via update_document() method", {
   docs <- mighty_documents(
-    file = test_path("test_study/documents.yml")
+    file = test_path("test_study/_documents.yml")
   )
 
   docs <- update_document(docs, id = "SUPPDOC001", title = "Updated")
@@ -46,7 +46,7 @@ test_that("document title can be correctly updated via update_document() method"
 
 test_that("document can be removed via remove_documents()", {
   docs <- mighty_documents(
-    file = test_path("test_study/documents.yml")
+    file = test_path("test_study/_documents.yml")
   )
 
   docs <- remove_documents(docs, id = "COMMENT001")
@@ -55,7 +55,7 @@ test_that("document can be removed via remove_documents()", {
   expect_equal(list_documents(docs),  c("SUPPDOC001", "METHOD001"))
 })
 
-test_that("mighty_study() reads documents.yml", {
+test_that("mighty_study() reads _documents.yml", {
   study <- mighty_study(test_path("test_study"))
 
   expect_s7_class(study@documents, mighty_documents)